+++ .eslintrc.cjs
@@ -37,7 +37,7 @@ module.exports = {
     {
       files: ['*.test.js'],
       rules: {
-        'no-unused-expressions': 'off',
+        '@typescript-eslint/no-unused-expressions': 'off',
       }
     }
   ],